Added upload of truncated versions

+++ tools/gulptasks/dashboards/dist-upload.js
@@ -7,6 +7,8 @@ const fs = require('fs');
 const gulp = require('gulp');
 const path = require('path').posix;
 
+const { dryrun } = require('yargs').argv;
+
 
 /* *
  *
@@ -155,22 +157,24 @@ async function uploadFile(
         path.relative(sourceFolder, sourceFile)
     );
 
-    await targetStorage.putObject({
-        Bucket: targetBucket,
-        Key: filePath,
-        Body: fileContent,
-        ACL: 'public-read',
-        ContentType: MIME_TYPE[path.extname(filePath)],
-        ...(!maxAge ? {} : {
-            CacheControl: `public, max-age=${maxAge}`,
-            Expires: new Date(NOW + (maxAge * 1000))
-        }),
-        ...(!isGzip ? {} : {
-            ContentEncoding: 'gzip'
-        })
-    });
+    if (!dryrun) {
+        await targetStorage.putObject({
+            Bucket: targetBucket,
+            Key: filePath,
+            Body: fileContent,
+            ACL: 'public-read',
+            ContentType: MIME_TYPE[path.extname(filePath)],
+            ...(!maxAge ? {} : {
+                CacheControl: `public, max-age=${maxAge}`,
+                Expires: new Date(NOW + (maxAge * 1000))
+            }),
+            ...(!isGzip ? {} : {
+                ContentEncoding: 'gzip'
+            })
+        });
+    }
 
-    logLib.message(filePath, 'uploaded');
+    logLib.message(filePath, 'uploaded', dryrun ? '(dryrun)' : '');
 }
 
 
@@ -288,6 +292,22 @@ async function uploadZips(
  *
  * */
 
+/**
+ * Returns truncated versions
+ * @param {string} fullVersion the full semver string
+ *
+ * @return {[string, string]}
+ * Array containing truncated versions ordered from most to least specific, i.e. ["1.1", "1"]
+ */
+function getTruncatedVersions(fullVersion) {
+    const [major, minor] = fullVersion.split('.');
+
+    return [
+        `${major}.${minor}`,
+        major
+    ];
+}
+
 
 /**
  * Uploads distribution files.
@@ -355,19 +375,28 @@ async function distUpload() {
         '.'
     );
 
-    const cdnVersionFolder = path.join(cdnFolder, release, '/');
+    // Upload versioned paths
+    const versions = [
+        release,
+        ...getTruncatedVersions(release)
+    ];
 
+    for (const version of versions) {
+        const cdnVersionFolder = path.join(cdnFolder, version, '/');
+
+        logLib.warn(`Uploading to ${cdnVersionFolder}...`);
+        await uploadFolder(
+            sourceFolder,
+            targetStorage,
+            bucket,
+            cdnVersionFolder,
+            HTTP_MAX_AGE.fiveYears
+        );
+    }
+
+    // Upload to path without version
     logLib.warn(`Uploading to ${cdnFolder}...`);
     await uploadFolder(sourceFolder, targetStorage, bucket, cdnFolder);
-
-    logLib.warn(`Uploading to ${cdnVersionFolder}...`);
-    await uploadFolder(
-        sourceFolder,
-        targetStorage,
-        bucket,
-        cdnVersionFolder,
-        HTTP_MAX_AGE.fiveYears
-    );
 
     // Hack
     const sourceCssFolder = path.join(sourceFolder, 'css');
